Restore ability of PACK to store raised errors
The feature was dropped because it wasn't thought to be used, but it
turns out that it was in the MAXMATCH tests.

It requires a bit of a hack to get past an assert regarding ovewriting
raised errors.  That assert may need to be dropped, but the workaround
helps this situation for now.