Workarounds for ACTION OF usages
The `...` operator and MATCH+ test were using ACTION OF, which is
now no longer relevant since frames are actions.

This works around the usages.  Plain PARAMETERS OF now works on
frames, and instead of ACTION OF BINDING OF 'RETURN this just
stores the block-combinator in a variable so it knows its own
identity another way.