Don't allocate 2GB buffers in ODBC on BLOB use
The column_size field in the bind columns--at least for firebird--was
reporting 2GB.

Allocating a fixed-size 2GB buffer is not a good approach.  Instead, we
pass in a 0 size buffer to SQLGetData, retrieve the actual length, and
then make another call with an allocated buffer.

As an added benefit, that buffer can be directly reclaimed as binary.