Add definitional behavior to BREAK and CONTINUE
This simple code was tried:

    for-each file ... [
        parse file [thru [".reb" | ".r"] <end> | (continue)]
        ...
    ]

But it did not work, because CONTINUE is not definitionally connected
to the loop it's inside.  This means the CONTINUE was actually doing
a continue of some arbitrary loop internal to the PARSE implementation
(e.g. looping across the rules in the combinator).

To deal with this, CONTINUE and BREAK (and STOP) are made "definitional"
e.g. they encode the frame of the loop they want to connect to in
the value cell of a variable named CONTINUE or BREAK or STOP that the
loop creates when it is instantiated.  References in the loop bind to
that variable instead of some global CONTINUE or BREAK.

This also led to revisiting the legality of storing non-managed
bindings in cells.  An exception had been made at one time for RETURN
because it lived as long as the frame, and it wasn't wanted for things
like natives to manage the array backing their arguments because it
generally shouldn't have to be managed.  This exception was confusing
and a bit dangerous, and no longer needed--because RETURN cells are
not put in natives.  Eliminating it simplifies the model.