Fix trash cell checking
Checking a trash cell used Cell_Heart(), which goes through the
READABLE() function.  It shouldn't be used for trash cells, so it
should not have worked.

The GC just wasn't seeing trash cells, until loop constructs began
managing their varlists to be put in BREAK and CONTINUE.  Then the
`iterator` variable could be visible in a trash state.