Bind refinement stack words to action, not frame
When refinements are pushed to the stack during action invocation,
they are pushed as words...and during the processing when it is
discovered which index in the frame they correspond to, that is
remembered by binding the word at that index...so the refinement
argument can be revisited directly at the right moment during
argument "pickups".

For the index to be valid, however, there needs to be a binding
supplied.  It was using the "varlist" of the frame being invoked,
but there was an issue that this varlist may not be managed...
meaning a cell was holding a binding to a non-managed thing.  That
is being made illegal.

This simply makes the words bound relative to the action that the
varlist was pushed for, instead of to the varlist itself.  That
action is managed, so it doesn't break the rule.