Use BindingHolder in C++ build to check binding
The BINDING() accessor was a shorthand for accessing the binding field
of the extra union in a cell, which also cast it away from being a
const Node*.  (The GC processes const Node* and will decay them to
a canon INACCESSIBLE series if the inaccessible flag gets set, and
that decay is done invasively on const Node*)

However, this casting made it not legal to use on the left hand side
of an assignment, so there had to be a `mutable_BINDING()` variant
to use to target the binding as an "lvalue".  Also, in order to
catch initializations of the binding to unmanaged nodes in the
debug build, there had to be an INIT_BINDING() function.

This moves checking to the C++ build, and eliminates the need for
a mutable_BINDING() or INIT_BINDING() by making BINDING() return a
holder class that operates on cell references, and can decide
whether to act like an lvalue or rvalue based on usage.