Make BLOCKIFY etc. empty blocks on void, not null
The concept behind NULL is that it's supposed to be testable easily
without generating word-fetch errors, but not be accepted by functions
that aren't doing tests for logic.  So BLOCKIFY and ENBLOCK etc.
shouldn't be willing to give an empty block on null.  If the feature
of producing an empty block is deemed useful at all, it should be a
reaction to void (as in APPEND DATA VOID being a no-op)