Global revisit of ANY-VALUE! in function specs
The ANY-VALUE! typeset once excluded nulls and voids, but today
those are legal variable states and hence legal in ANY-VALUE.  This
goes through and does some rethinking to when a parameter intended
actually any value vs. something that should only be an array element.
`<opt>` and `<void>` tags are subsumed by ANY-VALUE where appropriate.

It also changes ANY-VALUE! to the function ANY-VALUE?.  It's now a
type constraint and doesn't include unstable ISOTOPE! values, so it
could not be done with a conventional typeset.  This also permits the
recognition of the ANY-VALUE? intrinsic in the Init_Parameter()
function so it can be optimized as a single bit and avoid needing
to do word lookups in the parameter block.  The type checking savings
is significant, to about as much as 5% of runtime, for just this
change (and ANY-ATOM? done the same)