FOR-EACH-DATATYPE definitional continue hack
The FOR-EACH-DATATYPE loop-like construct in bootstrap is implemented
using PARSE.  So when it runs the body it is given, it's done with DO.

That previously ran the body in a simulated loop to get CONTINUE.  But
the simulated loop no longer finds the CONTINUE in `do body` because
it's not directly bound.

This sort of thing needs a systemic answer.  But as a quick hack, this
uses REPEAT 1 BODY to bind BREAK and CONTINUE each time the body runs.