Pass PROPERCASE-OF in to EMIT for binding
CSCAPE has built in abilities to turn things into all-uppercase or
all-lowercase.  However, there's no built in ability of propercase.
When that was added it had to be called in the escaped portion.

However, using a new executable for bootstrap, this won't be found
automatically as it's not in lib.  This is part of a more general
issue of "string interpolation is too hard", which is under review
as binding is revisited.