Kill off REBYMD struct
The year/month/date/zone of a DATE! is encoded in the Extra() payload
of the cell.  There was a separate datatype capturing just that part
of the date, which could be extracted and manipulated separately
from the date.

This creates an extra type--as if there should be routines that need
to operate on this type outside of a cell.  And it ties the hands of
the implementation (e.g. to need to store the time zone in the date
portion, when it's really part of the time portion).

Passing around a cell as currency is better.