Minor clarifying speedup: move bindinfo flag to symbol
A flag indicating whether a symbol was "in the binding table" was held
on the stub you'd reach after following a misc.  This moves the flag
to the canon symbol itself.

It saves a surprisingly negligible amount of CPU cycles (15,610
instructions to just boot and quit).  But it is certainly clearer than
the usage of the generic SERIES_FLAG_BLACK.