Kill "void skips while body" idea
Due to the loss of being able to CONTINUE in a condition, I thought it might
be interesting if someone could use VOID to ask to skip the body.

This would be in service of people putting lots of weird code in a while
condition, things that look like:

    while [
       switch do-something [
          ... [... true]
          ... [... true]
       ]
     ][
        ...
     ]

The idea being that instead of reducing it to a binary of truthy for
"run the body" falsey for "break the loop", you'd have a third option
of void for "skip the body and run the condition code again".

It was included speculatively merely because you could have done it
before with a CONTINUE, and has never been used.  I think the potential
for leading to a confusing situation is much higher than being truly
useful, and since WHILE is being used as a study in constrasts vs.
R3-Alpha and Red for implementation on the forum, this drops that.