Avoid over-allocation in release builds for poison tails
At one time, all arrays had "end cells" at their tail.  This information
was redundant with the length, and so it was eliminated--instead having
enumeration compute the tail pointer and stop when it was reached.

This lost a potential debugging advantage that end cells had been made
"mean" and would not permit being read...which helped catch some cases
of enumerating beyond the end of an array.  To make an even better
feature for debugging, the debug build added DEBUG_POISON_SERIES_TAILS
to prevent both reading and writing at the tail...and made allocations
an extra cell bigger.

Release builds weren't supposed to allocate space for the poison tail,
but pushed frames were doing so.  This corrects that.