Change "trash" to "unreadable" for debug build
The term "Trash" is being reclaimed for the new name of the unset
state (isotopic void, e.g. ~ isotope).  This means previous usages
of trash to mean a cell that is GC-safe but write-only need changing.

This makes it just called 'an unreadable'.  It could also be called
a "WriteOnly" but that feels a bit more convoluted.