Rename unset state from "NONE" to "TRASH"
Void isotopes are the choice for what to represent the state of
an unset variable.  This works well, but ISOTOPIC-VOID is not a
workable name for the state.

Calling the state "unset" is a hard veto, e.g. this is bad:

    >> unset? ~
    == ~true~  ; isotope

It is *variables* that are unset, not the state itself:

    >> x: ~

    >> unset? 'x
    == ~true~  ; isotope

(It might be more appropriate to call them UNDEFINED? since they are
actually set...just set to a state marking they have an uninteresting
value--but a value nonetheless.)

Refusing to call the state "unset" left questions about what the
alternatives would be.  Few palatable short names were available, and
the idea of calling it "trash" was overlooked because it seemed too
weird to have function specs say things like `return: <trash>` or
return statements write `return trash`.

The name "none" was available, and despite being suboptimal it
was chosen for the acceptability of `return: <none>` and `return none`.
That shallow rationale didn't lead to a good outcome, because of the
similiarities between none and null as words, and a lack of any
easy link to make between none and unsetness.

This led to a review of the concept of calling a ~ isotope "trash".
An observation making it seem more palatable was that since it has a
convenient literal form, there's not a reason to stress new users
with appearances of the name "trash" because code can just say things
like `return: [~]` or `return ~`.  The name becomes relevant only in
sophisticated code which actually has to manipulate the state more
invasively, at which time "trash" is a fine name if you choose to
use it (though you could again avoid it, e.g `^var = ~`