Get rid of nihil exemption for RETURN
Due to the commonness of returning `<trash>`, it once had a special
tag to allow you to omit a RETURN statement and just get trash.  This
was extended to `<nihil>`, partly as a legacy concern from a time
when the invisible state could not be expressed as a return statement.

The exception for nihil is removed, so you are now expected to do
`return nihil` and mark the function as `return: [nihil?]` in the
same way you'd do for anything else.