Make series accessibility more legible
Due to wanting the flags passed into series creation to have
the most common default, the flag for series accessibility wasn't
positive in the sense of SERIES_FLAG_ACCESSIBLE, or every call would
have to pass it in.  Instead, SERIES_FLAG_INACCESSIBLE is the odd
one out that gets added after the fact.

This meant to see if a series was accessible you'd have to write
Not_Series_Flag(s, SERIES_FLAG_INACCESSIBLE), and the double
negative was a bit annoying.

This makes some macros to help make things shorter and clearer, without
the double negatives.