Rethink "Inaccessible Series" to never be Series*
The idea of having the data backing a series be able to become not
accessible originated with the FFI.  It introduced a wrinkle that
when a series depended on external memory, then trying to read
out of it would trigger an error.

It became necessary to be able to make the backing data behind some
series "go away" for other reasons.  So Ren-C attempted to formalize
this a bit by saying it was the responsibility of cell extraction
functions to check if the series they were extracting were accessible
and if so, raise an error instead of giving back an array/series/string

However, it was still operating on pointers like Series* or Context*
and having to ask if pointers of those types happened to carry the
bit saying they were inaccessible.  This prevented rigorous checks
that any pointer which actually made it to the point of being cast to
a Series* or Context* etc. would be accessible--and DEBUG_CHECK_CASTS
actually had to make exceptions for it.

There was a particularly annoying example of this in the ENCLOSE and
CHAIN cases which was making frame varlists invalid even when there
were outstanding references to them as Array.  This wasn't necessary
for the CHAIN case, but ENCLOSE needed revisiting.  After changing
ENCLOSE to reuse the level it started with to do the invocation of
the outer function, there's no longer a need to manage corrupt
varlists while Action_Executor() is in effect.