RETURN/RUN for more ordinary-looking tail calls
This adds a new capability to call arbitrary functions and have
them reuse a Level* in the level stack.

It also gives a new syntax for REDO that may be more palatable
to some, to say `return/run @redo` instead of `redo` parameterized
with some local in the frame.

The old REDO functionality is still available.