Make PARAMETER! for unspecialized slots canon
There was an optimization that specializations would have an internal
parameter type in unspecialized slots, because the space was available
since there was no specialization.  This optimization gets harder to
hide from the user when trying to avoid transformations that would
fill the parameter slots with trash.

This makes a change which shifts to where parameters being in the
slots is a user-exposed feature, as the API for exposing reflection
and customization abilities for frames.