Move to param-and-key only function concept
Previously help information was encoded in objects, because the
paramlist only encoded typesets (not spec blocks) and did not
have room for strings.  Parameters now fully preserve spec
blocks so that information is redundant, and the description
strings can be embedded into the parameter.

This starts the transition by paring back the messy code for
making functions a bit.