Use MISC() of parameter array for optimization bytes
The parameter array is not currently using its MISC() or LINK()

In order to free space in the parameter cell for the help text, this
migrates the optimization bytes that were being stored in the cell
to the array.  It makes sense, because an unconstrained parameter
would not need these optimization bytes.  (If it did, it could
unset the node flag on the place where the array is written, and
write the bytes there...but that's not necessary.)