Rework isotope terminology to use "antiform"
Referring to the state of something like a WORD! when it could not be put in a
block used the term "word isotope".

This is not consistent with the idea that isotopes refer to variations of a
form which include the form itself.  Whatever this alternate form of WORD! is,
it's an isotope... but a non-marked WORD! is another isotope (plain isotope).

To try and be more consistent, this changes it so that "antiform" is the term
used when referring to the specific less-friendly isotope form, while saying
that in general the other forms (quasi and plain) are also isotopes.