Use HAS instead of IN for word-in-context-else-null
New applications of the IN function target the idea that if a word passed
is already bound, then IN will not add a binding to it.  A historical use
of IN would add a binding to words if found in the context, otherwise
return NULL.

This shifts to HAS as the function for the word lookup.