Rename IN to INSIDE, minimize noisy uses of INSIDE
An upcoming proposal would use the IN operator in a lot of places.

Saying IN is a shorthand for INSIDE permits the use of INSIDE to help see
the references while things are still being hammered out.

That's made easier if INSIDE is used sparingly.  Since in verbiage it's usually
just a lengthy way of saying IN, this touches up places that use the word.