Eliminate locals-gathering FUNCTION
An idea that originated in R3-Alpha to avoid needing to put /LOCAL
definitions in functions was to look for all the SET-WORD!s in the body
and turn those into locals.  This was a convenient idea but it had
side-effects of generating locals for SET-WORD!s that did not need
a local (e.g. from object members or dialects).

Ren-C committed to the idea of being able to do LET definitions.