Change GRAB to take variable with no quote
This was needed for the first cut at virtual binding, where QUOTED! will
eval in a way that doesn't have a binding.  But it's probably better
anyway--what it was trying to do was to clue you at the callsite that
GRAB knew the name/identity of the variable so it could actually update
the position of it (e.g. it's an IN/OUT parameter).  But that wasn't
really conveyed, and you'd have to know how grab worked to realize it.
So as long as you have to know how it works, you might as well know
that it takes the argument as a variable.