Hack to workaround stale L_current in SET-WORD!
The concept of lookback being stored in feeds is silly, now that cell
copying is fast (no derelativization needed).  Model should be simplified,
but current case has a bug...right hand evaluations can overwrite the
lookback buffer in some cases (sum: me + 1 is an example).  Work around
it pending a complete review of the issue.