Don't update L->feed.gotten on REB_FRAME
The L->feed.gotten optimization is supposed to help avoid redundant lookups
when L->feed.p points at a word type.  For some reason, it was being set to
a frame if a frame was seen literally in the stream of evaluation, and if a
GC happened at that moment it would fail an assert because it would double
check the word looked up to what it was supposed to...but it wasn't a word.