Narrow usages of attachment binding
The INDEX_ATTACHED state was being used sometimes when the module variable was
known to exist, which is slow.  This tightens that up, and changes it so that
when BINDING() is used as the left hand side of an assignment it will do some
validity checks.  This replaces checks in INIT_VAL_WORD_BINDING() and needs
the index and symbol to be assigned before the binding assignment.

(Still all very messy and experimental, but about to be changed a lot for
another experiment, so not stressing over that yet.)