Pure Virtual Binding Take II
This is a second try at 2001's attempt to collect complete environment
binding information on blocks.

A key difference in this iteration is that structural operations (like
PICK and FOR-EACH) do not attempt to propagate bindings.  Callers of
structural operations should tend to expect unbound results most of
the time, and choose a context in which to bind--where the block that
you picked from is one of the likely choices.