Streamline Cscape/Emit, prep for working interpolation
The coming changes to binding will allow for a BLOCK! to carry enough of an
environment to do binding and lookup for the accrued state.  This means
that a string-in-a-block can have enough information for interpolation.

However, in bootstrap it's still necessary to pass in words that are
needed for the interpolation.  This used to be done via a /WITH refinement
to CSCAPE that was a variadic skippable BLOCK! for emit.  This simplifies
that by just taking a BLOCK! with extra words in it... or a string.  The
extra words would not be used by newer interpreters, and could someday
be eliminated.