Cache whether words are in function or not in body
Function bodies are now generally conceptually unbound.  This would suggest
we have to pay for environment lookup of words on every run.  However, we can
use the binding space in the unbound words to hold information about whether
a word can be found in frames that are instances of that function or not.

This avoids needing to search the function frame manually for these cases.