Introduce Element(*) type for non-isotope, non-void cells
The "Cell" type was doing double-duty as a way of saying that something was
an array element that hadn't been connected with a binding, and that
it shouldn't be initialized with an isotope.

Changes in binding are throwing a lot of that out the window.  This makes
a new type specifically to capture the idea of array elements.