Use blank instead of void to opt out of set-group assignment
The "current" cell is always supposed to be something that can be in source,
so having it be void breaks that invariant.  This sets it to blank when a
set-group evaluates to void and jumps up to use the SET-WORD! code.

There are probably better ways to do this, but this is just to keep it
compiling under new rules.