Stop Is_Void() and Is_Logic() tests on Element*
This ripples through changes you get when you stop making it possible to
test elements for void or logic.

It accomplishes this by means of a new Need() wrapper class which is
made as a synonym for Sink(), to turn the inheritance model on its head
so that more derived classes like Element won't automatically convert
to the base type.