Move_Cell() and Copy_Cell() protect Elements from Atoms/Values
This makes casual uses of move or copy of cells not able to put unstable
isotopes into Value cells, or any antiforms into Element cells.

It requires a kind of ugly template trick to remove an overload from the
candidate set, but I don't see any better way to do it.