Get rid of NoQuote(Cell*)
This shifts it so that Cell* is the base type of which you can't ask what its
type is.  Many callsites were changed to more clearly use the Element or Atom
or Value subclass to help indicate what kinds of antiforms they could hold,
though the choices aren't perfect and need review.