Make array stepping able to reuse evaluator level
Killing off the stale bit meant that a separate cell had to be used as
a temporary after each evaluation, to make sure that invisible products
like NIHIL or BARRIER would not overwrite the previous result.  Rather
than put an extra temporary cell in every frame, a separate executor
was created with the responsibility of doing the steps in an array.
However, this was a large amount of weight for something that only
needed a single cell's worth of data.

This rethinks the idea to put the cell on the data stack, and the
stepper executor is able to actually pass through to the evaluator
executor.  It can also create a reified sublevel if there's a good
reason to do that (e.g. if debuggers want reified levels to get hooks
into each completion step of the sub-evaluator).

A speedup of about 1.5% was observed on a plain startup and shutdown,
but would represent more of a speedup on usermode code.

As long as this was being done, the Array_Executor() was renamed
to Stepper_Executor() and put in its own file.