Handle new semantics for IMPORT @WORD
Previously IMPORT @WORD had distinct meaning from IMPORT 'WORD but now
these two are the same.  This isn't great (and also not great that
DO 'WORD will run a script from a module repository) but changing
to accept it so that things can keep running for now.

Likely DO of BLOCK! will not be used as a synonym of eval, and
IMPORT [@WORD] and DO [@WORD] will be used.