Reorder %types.r to put all word/array types together
The original ordering of %types.r was done in order to put all the inert
types together.  So all the THE-XXX! types were together, the basic
BLOCK! was at the head of the plain types and then all the plain types
together (so block could be inert), and all the SET-XXX! and GET-XXX!
types were together etc.

This has turned out to be a pretty uninteresting optimization, because
the whole idea of what's inert has been turned on its head.  Blocks
aren't inert anymore, and nowadays there's arbitrary levels of quoting
which kind of shows that there's more "trivial evaluations" and
"complex evaluations".

This puts things together in a more interesting way that can be
exploited by optimizations (but doesn't add the optimizations yet,
it's actually a slight micro bit slower with *just* this change).