Simplify deep copying to just a boolean
There was an idea in historical Rebol that you could specify to copying
what things to copy deeply, and this could include things like objects...but
then there were cases that would be bad to copy, like PORT!, which were
excluded.

The idea of being able to copy an object is a pretty large semantic
minefield--there's little guarantee that this can be done automatically in
any meaningful way.  Deep copying should likely be limited to block and
series structure--which is mostly what it was doing.

This makes that the only kind of deep copying, and the system still
appears to run with this homogeneous rule.  It's likely better to build on
this for now than the confusing situation that was in place (also breaks
a dependency on 64-bit typesets which was limiting the type count).