ANY-VALUE? and ELEMENT? Exceptions Less Exceptional
These two functions don't fit the pattern of the rest of the typeset
functions, but they were slipped into it with the idea that they needed
to make predicate values for boot.  All boot uses are ANY-VALUE? and
ELEMENT? directly in the parameters (after fixing an issue).

This is a step being taken toward trying to simplify the typecheckers
in general.