Switch typechecking intrinsics to use "Decider" functions
This makes it so that the intrinsics will use whatever implementation backs
the functions like Any_Array()... which may be typeset based or be range
based or other optimizations.

It's a step toward increasing the optimization for typechecking, as well
as a way toward escaping the 64-bit bitset limit caused by TS_XXX, so that
the full 256 types can be used for the heart byte.