Kill off Did_Inert_Optimize_Complete()
Out of curiosity I deleted the calls to this and checked the result
in callgrind to see how much it was saving.  *It was making things
actually slower.*  Not only that, it's broken in the current model
because the inert behaviors didn't quite match what the evaluator
executor was doing.

Removed to revisit the question of how to do an actually effective
optimization.