Make TYPE OF NULL raise an error
The general question of TYPE OF antiforms is up in the air, but not the
idea that TYPE OF NULL should be an error.

If you say TYPE OF MAYBE NULL that's TYPE OF VOID, which obeys the void
in null out error.  Which isn't to say that void is null, but rather
that it has no type.