Make sure only Heart passed to places setting HEART_BYTE()
The upcoming idea with separating Heart from the Kind type is to help avoid
accidentally getting "pseudotypes" that are supposed to be implied by quote
bytes from winding up in the heart byte slot on accident.  It also will
prevent the need from having to test for quoting bytes when they're not
in the domain of consideration, hopefully making some things faster.

This moves a little further toward the separation (though Heart and Kind
are still the same type at this point).