Fix rare assert in Shuffle_Array()
Shuffle_Array() did something strange, which was to declare a raw Element
object vs. using DECLARE_ELEMENT().  There is a destructor which runs and
checks to make sure all constructed Value/Element/Atom are either 0 in
their cell mask at time of destruction or have some initialized content.
This was failing that test.

Also removes a weird not-worth-it cell "blitting" optimization, but adds
a sanity check that a blit is legal w.r.t. CELL_MASK_PERSIST.