Flip implementation from 64 type limit to 64 typeset limit
The original typeset implementation used 64-bit integers that were OR'd together
flags of all the types in each typeset.  This meant requiring 64-bit integers in
the compiler (which is actually a design question under review) but also was the
last restriction limiting the system to 64 fundamental types.

This switches it around, instead using a global array the length of the number
of types, where each integer in the array is OR'd together flags of the typesets
that type is a member of.  This switches the limitation so it's on the number
of typesets instead of the number of types.  Right now it uses uint_fast32_t
for the typesets so there's at most 32, but it could be 64.  (We're not in danger
of hitting the 32 builtin typeset limit at this time.)