Stop Get_Lex_Class(b) == LEX_DELIMIT Style Bugs
When adding the $ type to the scanner, I accidentally tested against
LEX_DELIMIT instead of LEX_CLASS_DELIMIT.  It was not caught by the compiler
because it was the comparison of an enum with a plain integer.  This wraps
the LEX_DELIMIT flags in an enum that is type incompatible when enum
comparison warnings are enabled to stop that in the future.