Solve bug in WORD! aliased TEXT! by simplifying code
There's a lot of simplification that needs to be done of the ANY-UTF8!
types.  This just does a little bit, to address an issue that popped
up where if you did AS TEXT! of a WORD! and then weren't at the
head of that text, incorrect length and offset calculations were
being done.

The way to clean this up is to make "stringlike" things more uniform
by just glossing over the little differences (e.g. act like the index
is zero for a WORD! vs. force you to use a different code path that
doesn't talk about indices at all).