Make the ANY-XXX! type constraints use &any-xxx?
Now that type constraints are easily written out as just TYPE-WORD!, the
concern over wanting to distinguish ANY-SERIES?! from ANY-SERIES! that can't
be used e.g. as `any-series! = type of value` goes away.  Because

    parse [1 20.2] [some &any-number?]

Is better than:

    parse [1 20.2] [some any-number?!]

It's more generalized, uses fewer definitions, requires one less lookup, and
has a clear rationale for the decoration.

It's still up in the air if things like LOGIC! and CHAR! are going to be
things you can say `logic! = type of true` with, by generating some kind of
a TYPE-BLOCK!.  So those aren't changed for now.

Also retrofits the bootstrap shim so it can use things like &any-series? in
PARSE too, by making &any-series? a synonym for any-series! since it's legal
to start words with ampersands.