Redesign %import-shim.r to use --import
In order to preprocess the source code for fundamentally new syntax, the DO
process itself must be hooked vs. something that happens after a script has
already been loaded.  This rethinks the import shim to take advantage of an
old feature in the interpreter of --import which runs IMPORT on a module
before anything runs.

Unfortunately, this trips over bugs in the existing bootstrap executable,
making it seem like it's time to bite the bullet and make a new one.  This
adjusts the process, but requires one of those new executables.