Remove order requirement for ANY-INERT?
There were too many constraints, and making inertness indicated in the types.r
table by being in ANY-INERT? is a sane way to convey this requirement.

Switch it to a bit on the cell if it turns out that knowing it fast is
important (an Any_Inert() option was recently ripped out because it was
actually making things slower).