Drop superfluous set-block for |> and <|
The idea of creating a second-class citizen for words made `|>` and `<|`
into forms that could not have sigils for a time.  Now they merely can't
be put in paths or sequences, but they can have sigils.

This means special treatment based on them not being able to be set-words
in bootstrap code is not needed.