Add New SIGIL! Datatype
Having certain words that couldn't have sigils creates problems (e.g.
MAKE OBJECT! with a key that's named ":" doesn't know how to generate
a SET-WORD! for that case, it's just the tip of the iceberg.)

However, we do want to give ^ $ & : @ evaluator behavior and allow
them in dialects.

This adds a new SIGIL! type to capture just these cases.  It also
gives meaning to :: as the sigil of a SET-WORD!.

The cell format for SIGIL! is the same as ISSUE! (plus a cache of the
sigil enum value) so it has the UTF-8 representation in the cell, and
can use common code for handling rendering.

As a first step, this doesn't replace the word handling for the cases
that were relying on function calls in the evaluator.