Eliminate sigil-WORD! exceptions, add sigil eval behaviors
This kills off the words that could not be sigil'd, fully replacing them
with the SIGIL! instances.

`^` is a synonym for META
`@` is a tweaked version of THE (tweak to be revisited)
`$` is a binding operator equivalent to IN []

For the sake of having it do something, `&` does TRY TYPE OF.  Not that
you would want to use it in source (hopefully) but generated code might
find an advantage in performance...or, something.

`::` and `:` aren't given any definition yet.