Correct improper use of Is_Antiform_Unstable()
The fail() macro calls a function that will synthesize a BAD-VALUE error
if it sniffs the pointer it receives and finds it to be a cell.

When this was happening, there was an assertion that *meant* to say
assert(Is_Stable()) but instead said assert(not Is_Antiform_Unstable()).
That test is only supposed to be used on things you are already sure
are antiforms.

Simple to fix.  However, it exposed the fact that there are no tests that
were triggering that, including no test which tries to use an invalid
type predicate name.  This adds such a test, so now there's at least one.