Make `<end>` UPARSE combinator invisible
Most cases using `<end>` or `to <end>` or `thru <end>` have no interest in
the value of the tail of the series.

If the tail is desired, that can be done with `<end> <here>`

Making `<end>` invisible means you can avoid writing `elide <end>` in the
much more common cases.