Eliminate superfluous DID ALL in tests
There is an idea behind the tests that each test must evaluate to a logic.
Since ALL returns either null or the arbitrary last truthy value it evaluated,
it could return something "truthy" without passing the test.

A long time ago, this caused a reactionary replacement of every ALL with
DID ALL in the tests.

But the fact is that almost every one of these tests has a last item in the
return list which is a logic.  And there's nothing to be concerned about
with returning null instead of false in the case of failure, because null
fails the test just as well as false does!

This commit cleans off the cruft and leaves the DID ALL only in the rarest
situations where it actually is needed.