Fold in tests from old-parse-THE and old-parse-INTEGER
When the question of if integers would evaluate to themselves in
UPARSE was being considered, the existing tests that had been taken
from R3-Alpha and Red were put aside in a file not run by the
core-tests.  They were forgotten about.  This rewrites them in terms
of new repeat and moves them into the main parse repeat test file.

Three was another file for old-parse-THE which deals with an issue that
is still open, but this moves the content commented into the main
file just so it doesn't get forgotten about.