Give UPARSE definitional errors IDs
This gives IDs to the two errors currently being returned... parse-mismatch
and parse-incomplete.  The concept behind the errors returned by UPARSE is
that they could give you some kind of accounting log of what the errors
were, but that feature is not implemented yet so these are the only two
errors you get.

With the errors returned, many tests can be made more obvious that they
are failing and why by using the error ID notation in the test dialect.