Rename MATCH-PARSE to VALIDATE, make a combinator too
This fixes MATCH-PARSE (no test checked that it ever returned null,
and a PARSE change made it so it wouldn't), and adds a test for that.

Also renames MATCH-PARSE to VALIDATE (the parameters were in the wrong
order to compare it to MATCH) and makes it available as a combinator.

Some question is in the air of whether SUBPARSE should just be called
PARSE.  This may be revisited in the near future.  For now, it's still
called SUBPARSE.