Make DECLARE_XXX() macros for cells match subclasses
The DECLARE_LOCAL() macro needed additional forms to cover cases where
unstable antiforms were legal as well as where no antiforms were.
But because this predated the Value/Element/Atom subclasses, the names
weren't aligned with these types.

This brings that into alignment and updates the associated comment to
explain what these are actually for (and not for).